Move desk drop actions into widget handlers
