+++ apps/ui/src/ui-desks/widgets/post/definition.ts
@@ -1,12 +1,17 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { post } from '@wordpress/icons';
+import { getSiteContentMediaDropActions } from '@/ui-desks/widget-actions/drop-handlers/site-content-media-actions';
 import { isMediaWidgetProps, MEDIA_WIDGET_TYPE } from '@/ui-desks/widgets/media/types';
 import {
 	PostWidgetComponent,
 	PostWidgetThumbnailComponent,
 } from '@/ui-desks/widgets/post/component';
 import { isPostWidgetProps, POST_WIDGET_TYPE, type PostWidget } from './types';
-import type { WidgetDefinition } from '@/ui-desks/widgets/types';
+import type {
+	WidgetCustomDropActionContext,
+	WidgetCustomDropActionIntent,
+	WidgetDefinition,
+} from '@/ui-desks/widgets/types';
 
 export const postWidgetDefinition = {
 	type: POST_WIDGET_TYPE,
@@ -57,6 +62,35 @@ export const postWidgetDefinition = {
 				sourceWidget.widgetProps.mediaId !== null &&
 				isPostWidgetProps( targetWidget.widgetProps ) &&
 				targetWidget.widgetProps.postId > 0,
+			getActions: getPostMediaDropActions,
 		},
 	],
 } satisfies WidgetDefinition< PostWidget >;
+
+function getPostMediaDropActions(
+	intent: WidgetCustomDropActionIntent,
+	context: WidgetCustomDropActionContext
+) {
+	const mediaProps = intent.sourceWidget.widgetProps;
+	const postProps = intent.targetWidget.widgetProps;
+	if (
+		! isMediaWidgetProps( mediaProps ) ||
+		mediaProps.mediaId === null ||
+		! isPostWidgetProps( postProps )
+	) {
+		return [];
+	}
+
+	return getSiteContentMediaDropActions( {
+		kind: 'post',
+		contentId: postProps.postId,
+		attachLabel: __( 'Attach to post' ),
+		media: {
+			id: mediaProps.mediaId,
+			url: mediaProps.url,
+			alt: mediaProps.alt,
+			kind: mediaProps.mediaKind,
+		},
+		context,
+	} );
+}
