+++ apps/ui/src/ui-desks/widgets/scratchpad/definition.ts
@@ -1,5 +1,19 @@
 import { __ } from '@wordpress/i18n';
 import { external } from '@wordpress/icons';
+import {
+	buildWidgetContextDisplayMessage,
+	buildWidgetContextPrompt,
+} from '@/ui-desks/chats/widget-context';
+import {
+	completeConnectorPreview,
+	createConnectorPreview,
+	toPlainPoint,
+	updateConnectorEnd,
+} from '@/ui-desks/connectors/editor-commands';
+import {
+	RECTANGLE_WIDGET_SHAPE_TYPE,
+	type RectangleWidgetShape,
+} from '@/ui-desks/shapes/rectangle-widget/types';
 import { isMediaWidgetProps, MEDIA_WIDGET_TYPE } from '@/ui-desks/widgets/media/types';
 import {
 	ScratchpadWidgetComponent,
@@ -14,7 +28,14 @@ import {
 	isScratchpadWidgetProps,
 	type ScratchpadWidget,
 } from '@/ui-desks/widgets/scratchpad/types';
-import type { WidgetDefinition } from '@/ui-desks/widgets/types';
+import type {
+	WidgetCustomDropActionContext,
+	WidgetCustomDropActionIntent,
+	WidgetDefinition,
+} from '@/ui-desks/widgets/types';
+import type { Editor, JsonObject } from 'tldraw';
+
+const BUILD_SOMETHING_LIKE_THIS_PROMPT = __( 'Build something like this' );
 
 export const scratchpadWidgetDefinition = {
 	type: SCRATCHPAD_WIDGET_TYPE,
@@ -50,6 +71,93 @@ export const scratchpadWidgetDefinition = {
 			type: 'custom',
 			sourceTypes: [ MEDIA_WIDGET_TYPE ],
 			canHandle: ( sourceWidget ) => isMediaWidgetProps( sourceWidget.widgetProps ),
+			getActions: getScratchpadMediaDropActions,
 		},
 	],
 } satisfies WidgetDefinition< ScratchpadWidget >;
+
+function getScratchpadMediaDropActions(
+	intent: WidgetCustomDropActionIntent,
+	context: WidgetCustomDropActionContext
+) {
+	if (
+		! isMediaWidgetProps( intent.sourceWidget.widgetProps ) ||
+		! isScratchpadWidgetProps( intent.targetWidget.widgetProps )
+	) {
+		return [];
+	}
+
+	return [
+		{
+			label: BUILD_SOMETHING_LIKE_THIS_PROMPT,
+			onClick: () =>
+				context.runAction( async () => {
+					updateScratchpadReference( context.editor, intent );
+					await context.startChatWithPrompt( {
+						prompt: buildWidgetContextPrompt( BUILD_SOMETHING_LIKE_THIS_PROMPT, [
+							intent.sourceWidget,
+							intent.targetWidget,
+						] ),
+						displayMessage: buildWidgetContextDisplayMessage( BUILD_SOMETHING_LIKE_THIS_PROMPT, [
+							intent.sourceWidget,
+							intent.targetWidget,
+						] ),
+					} );
+				} ),
+		},
+		{
+			label: __( 'Use this image' ),
+			onClick: () =>
+				context.runAction( () => createConnectorBetweenWidgets( context.editor, intent ) ),
+		},
+	];
+}
+
+function updateScratchpadReference( editor: Editor, intent: WidgetCustomDropActionIntent ) {
+	const mediaProps = intent.sourceWidget.widgetProps;
+	const targetShape = editor.getShape( intent.targetShapeId ) as RectangleWidgetShape | undefined;
+	if (
+		! isMediaWidgetProps( mediaProps ) ||
+		! targetShape ||
+		targetShape.type !== RECTANGLE_WIDGET_SHAPE_TYPE ||
+		! isScratchpadWidgetProps( targetShape.props.widgetProps )
+	) {
+		return;
+	}
+
+	const currentDescription = targetShape.props.widgetProps.description ?? '';
+	editor.updateShape< RectangleWidgetShape >( {
+		id: targetShape.id,
+		type: RECTANGLE_WIDGET_SHAPE_TYPE,
+		props: {
+			widgetProps: {
+				...targetShape.props.widgetProps,
+				description: currentDescription.trim()
+					? currentDescription
+					: BUILD_SOMETHING_LIKE_THIS_PROMPT,
+				reference: {
+					mediaId: mediaProps.mediaId,
+					url: mediaProps.url,
+					alt: mediaProps.alt,
+				},
+			} satisfies JsonObject,
+		},
+	} );
+}
+
+function createConnectorBetweenWidgets( editor: Editor, intent: WidgetCustomDropActionIntent ) {
+	const sourceBounds = editor.getShapePageBounds( intent.sourceShapeId );
+	const targetBounds = editor.getShapePageBounds( intent.targetShapeId );
+	if ( ! sourceBounds || ! targetBounds ) {
+		return;
+	}
+
+	const connectorShapeId = createConnectorPreview(
+		editor,
+		intent.sourceShapeId,
+		toPlainPoint( sourceBounds.center ),
+		toPlainPoint( targetBounds.center )
+	);
+	completeConnectorPreview( editor, connectorShapeId, intent.targetShapeId );
+	updateConnectorEnd( editor, connectorShapeId, toPlainPoint( targetBounds.center ) );
+}
